refactor: use i64s for page size and index

+++ entity/src/entities/meta_last_page.rs
@@ -6,9 +6,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "meta_last_page")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub page_size: i32,
+    pub page_size: i64,
     #[sea_orm(primary_key, auto_increment = false)]
-    pub page_index: i32,
+    pub page_index: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
